utils.py
import httpx
import os
from dotenv import load_dotenv # type: ignore

load_dotenv()
GROQ_API_KEY = os.getenv("GROQ_API_KEY")

async def stream_groq_response(messages: list):
    url = "https://api.groq.com/openai/v1/chat/completions"
    headers = {"Authorization": f"Bearer {GROQ_API_KEY}"}
    payload = {
        "model": "llama-3.3-70b-versatile",  # or whichever Groq model
        "messages": messages,
        "stream": True
    }

    async with httpx.AsyncClient(timeout=None) as client:
        async with client.stream("POST", url, headers=headers, json=payload) as response:
            async for line in response.aiter_lines():
                if line.startswith("data: "):
                    if line == "data: [DONE]":
                        break
                    yield line[6:]
